# TravelGuideApi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Travel Guide API",
    "description": "API for the Travel Guide mobile application",
    "contact": {
      "name": "Travel Guide Team"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Destinations": {
      "get": {
        "tags": [
          "Destinations"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "SortOrder",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DestinationDtoPaginatedResponseApiResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Destinations"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DestinationCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DestinationDtoApiResponse"
            }
          }
        }
      }
    },
    "/api/Destinations/{id}": {
      "get": {
        "tags": [
          "Destinations"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DestinationDtoApiResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Destinations"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DestinationUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DestinationDtoApiResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Destinations"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BooleanApiResponse"
            }
          }
        }
      }
    },
    "/api/Destinations/search": {
      "get": {
        "tags": [
          "Destinations"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "type": "string",
            "default": "",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "region",
            "type": "string",
            "default": "",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "category",
            "type": "string",
            "default": "",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "minRating",
            "type": "number",
            "format": "float",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "SortOrder",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DestinationDtoPaginatedResponseApiResponse"
            }
          }
        }
      }
    },
    "/api/Destinations/popular": {
      "get": {
        "tags": [
          "Destinations"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "count",
            "type": "integer",
            "format": "int32",
            "default": 5,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DestinationDtoListApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BooleanApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "boolean"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "DestinationCreateDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "pointsOfInterest": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "imageUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "coverImageUrl": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "category": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DestinationDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "averageRating": {
          "format": "float",
          "type": "number"
        },
        "pointsOfInterest": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "imageUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "coverImageUrl": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DestinationDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/DestinationDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "DestinationDtoListApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DestinationDto"
          }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "DestinationDtoPaginatedResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DestinationDto"
          }
        },
        "totalCount": {
          "format": "int32",
          "type": "integer"
        },
        "pageNumber": {
          "format": "int32",
          "type": "integer"
        },
        "pageSize": {
          "format": "int32",
          "type": "integer"
        },
        "totalPages": {
          "format": "int32",
          "type": "integer"
        },
        "hasPrevious": {
          "type": "boolean"
        },
        "hasNext": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "DestinationDtoPaginatedResponseApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/DestinationDtoPaginatedResponse"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "DestinationUpdateDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "pointsOfInterest": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "imageUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "coverImageUrl": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "category": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}